wip v2 - refactor to TaskWorker
